fix var names

+++ .pipelines/EV2Specs/ServiceGroupRoot/Shell/Run/Run.ps1
@@ -29,41 +29,81 @@ try {
     Invoke-WebRequest -Uri $env:PWSH_PACKAGES_TARGZIP -OutFile packages.tar.gz
     Invoke-WebRequest -Uri $env:PMC_METADATA -OutFile pmcMetadata.json
 
-    $settingsFile = Join-Path '/package/unarchive/Run' -ChildPath 'settings.toml'
-    $settingsFileExists = Test-Path -Path $settingsFile
-    Write-Verbose -Verbose "settings.toml exists: $settingsFileExists"
+    # create variables to those paths and test them
+    $repoListFilePath = Join-Path -Path "/package/unarchive/" -ChildPath "repoList.json"
+    $repoListPathExists = Test-Path $repoListFilePath
+    if (!$repoListPathExists)
+    {
+        Write-Verbose -Verbose "repoList.json expected at $repoListFilePath does not exist"
+        return 1
+    }
+
+    $mappingFilePath = Join-Path "/package/unarchive/" -ChildPath "mapping.json"
+    $mappingFilePathExists = Test-Path $mappingFilePath
+    if (!$mappingFilePathExists)
+    {
+        Write-Verbose -Verbose "mapping.json expected at $mappingFilePath does not exist"
+        return 1
+    }
+
+    $packagesTarPath = Join-Path -Path "/package/unarchive/" -ChildPath "packages.tar.gz"
+    $packagesTarPathExists = Test-Path $packagesTarPath
+    if (!$packagesTarPathExists)
+    {
+        Write-Verbose -Verbose "packages.tar.gz expected at $packagesTarPath does not exist"
+        return 1
+    }
+
+    $metadataFilePath = Join-Path -Path "/package/unarchive/" -ChildPath "pmcMetadata.json"
+    $metadataFilePathExists = Test-Path $metadataFilePath
+    if (!$metadataFilePathExists)
+    {
+        Write-Verbose -Verbose "pmcMetadata.json expected at $metadataFilePath does not exist"
+        return 1
+    }
+
+    # files in the extracted Run dir
+    $configPath = Join-Path '/package/unarchive/Run' -ChildPath 'settings.toml'
+    $configPathExists = Test-Path -Path $configPath
+    if (!$configPathExists)
+    {
+        Write-Verbose -Verbose "settings.toml expected at $configPath does not exist"
+        return 1
+    }
+
     $pythonDlFolder = Join-Path '/package/unarchive/Run' -ChildPath 'python_dl'
     $pyPathExists = Test-Path -Path $pythonDlFolder
-    Write-Verbose -Verbose "python_dl folder path exists: $pyPathExists"
+    if (!$pyPathExists)
+    {
+        Write-Verbose -Verbose "python_dl expected at $pythonDlFolder does not exist"
+        return 1
+    }
 
-    Write-Verbose -Verbose "doing gci"
-    Get-ChildItem "/package" -Recurse
     Write-Verbose -Verbose "Installing pmc-cli"
-    crane version
-    # python.exe -m pip install --upgrade pip
     pip install --upgrade pip
     pip --version --verbose
     pip install /package/unarchive/Run/python_dl/*.whl
 
     Write-Verbose -Verbose "Test pmc-cli"
+    pmc -d -c $configPath repo list --accessible || exit 1
 
-    pmc -d -c $settingsFile repo list --accessible || exit 1
-
-    $metadataFilePath = Join-Path -Path "/package/unarchive/" -ChildPath "pmcMetadata.json"
+    # Get metadata
+    $channel = ""
+    $packageNames = @()
     $metadataContent = Get-Content -Path $metadataFilePath | ConvertFrom-Json
     $releaseVersion = $metadataContent.ReleaseTag
 
+    if ($releaseVersion.Contains('-')) {
+        $channel = 'preview'
+        $packageNames = @('powershell-preview')
+    }
+    else {
+        $channel = 'stable'
+        $packageNames = @('powershell')
+    }
 
-
-    # $repoListFilePath = Join-Path -Path "/package/unarchive/" -ChildPath "repoList.json"
-    # $repoListPathExists = Test-Path $repoListFilePath
-    # if (!$repoListPathExists)
-    # {
-    #     Write-Verbose -Verbose "repoList.json expected at $repoListFilePath does not exist"
-    #     return 1
-    # }
-
-    # $repoList = Get-RepoList -RepoConfigPath $configPath -OutFile $repoListPath #TODO: seems to require repoclient cli tool
+    Write-Verbose -Verbose "---Getting repository list---"
+    #$repoList = Get-RepoList -RepoConfigPath $configPath -OutFile $repoListPath #TODO: seems to require repoclient cli tool
 
     # # Get-PackageInfo()
     # $mappingFilePath = Join-Path -Path "/package/unarchive/" -ChildPath "mapping.json"
